Remove gemini-stream.ts
This module is no longer needed and the types can be provided from types.
