--- packages/cli/src/core/gemini-stream.ts
@@ -1,19 +1,0 @@
-/**
- * @license
- * Copyright 2025 Google LLC
- * SPDX-License-Identifier: Apache-2.0
- */
-
-// Only defining the state enum needed by the UI
-export enum StreamingState {
-  Idle = 'idle',
-  Responding = 'responding',
-  WaitingForConfirmation = 'waiting_for_confirmation',
-}
-
-// Copied from server/src/core/turn.ts for CLI usage
-export enum GeminiEventType {
-  Content = 'content',
-  ToolCallRequest = 'tool_call_request',
-  // Add other event types if the UI hook needs to handle them
-}
