+++ packages/cli/src/ui/App.tsx
@@ -6,7 +6,7 @@
 
 import React, { useState, useMemo } from 'react';
 import { Box, Text } from 'ink';
-import type { HistoryItem } from './types.js';
+import { StreamingState, type HistoryItem } from './types.js';
 import { useGeminiStream } from './hooks/useGeminiStream.js';
 import { useLoadingIndicator } from './hooks/useLoadingIndicator.js';
 import { useInputHistory } from './hooks/useInputHistory.js';
@@ -15,7 +15,6 @@ import { HistoryDisplay } from './components/HistoryDisplay.js';
 import { LoadingIndicator } from './components/LoadingIndicator.js';
 import { InputPrompt } from './components/InputPrompt.js';
 import { Footer } from './components/Footer.js';
-import { StreamingState } from '../core/gemini-stream.js';
 import { ITermDetectionWarning } from './utils/itermDetection.js';
 import {
   useStartupWarnings,
