+++ packages/cli/src/ui/types.ts
@@ -7,6 +7,20 @@
 import { ToolCallConfirmationDetails } from '@gemini-code/server';
 import { ToolResultDisplay } from '../tools/tools.js';
 
+// Only defining the state enum needed by the UI
+export enum StreamingState {
+  Idle = 'idle',
+  Responding = 'responding',
+  WaitingForConfirmation = 'waiting_for_confirmation',
+}
+
+// Copied from server/src/core/turn.ts for CLI usage
+export enum GeminiEventType {
+  Content = 'content',
+  ToolCallRequest = 'tool_call_request',
+  // Add other event types if the UI hook needs to handle them
+}
+
 export enum ToolCallStatus {
   Pending = 'Pending',
   Invoked = 'Invoked',
